renewables/wind/plant/simulation-replay/manager/requirements.txt
docopt==0.6.2
elasticsearch==8.3.1
numpy==1.23.1
requests==2.28.1
tqdm==4.64.1
python_dateutil==2.8.2